update cmc model

+++ train_cmc.py
@@ -94,9 +94,9 @@ class Workspace:
                     for _ in range(self.cfg.num_evaluations):
                         video_i, video_p, video_n = next(self.valid_dataloader_iter)
 
-                        video_i = video_i.to(device=utils.device())
-                        video_p = video_p.to(device=utils.device())
-                        video_n = video_n.to(device=utils.device())
+                        video_i = video_i.to(device=utils.device(), dtype=torch.float)
+                        video_p = video_p.to(device=utils.device(), dtype=torch.float)
+                        video_n = video_n.to(device=utils.device(), dtype=torch.float)
 
                         m, _ = self.encoder.evaluate(video_i, video_p, video_n)
